Proposal to authentification, aiming for JWT
This commit is mostly a temporary mechanism on clientside to pass credentials to server.
The server now needs to actually check the credentials (DB interface missing)
and sign a userid token (JWT) that is sent back to the client for future authentifications.

The client stores the recieved token in localstorage.